Update pipeline script to match new cog script
A change in `generate_cogs.sh` meant that the script only expects two inputs --
the name of the input tif and the number of bands. This commit updates the
workflow tool to reflect that. Note that as of this update, the input CSV
expects only two columns: the input tif file and the number of bands.