Allow input csv in subfolder
This wound up being a bit tricker than I anticipated, and I settled on an
an approach that wound up injecting an additional assumption into the workflow.
Namely, we must now assume that any CSV files used to start workflows have
unique BASE names. That is, if I have a CSV in the input bucket like:
`name/of/my/files-raw.csv`
I can't have any other `files-raw.csv` in the input bucket. So this is illegal:
`different/path/to/files-raw.csv`
But this is okay:
`different/path/to/different-raw.csv`